requirements.txt
streamlit
pillow
requests
openai
transformers
beautifulsoup4
diskcache
torch
huggingface_hub
accelerate